OOP - Homework 3 - Task 2

+++ OOP/OOP_Principles_P2/Task2/IDeposit.cs
@@ -4,6 +4,6 @@
 
     public interface IDeposit
     {
-        void MakeDeposit(double depositAmount);
+        void MakeDeposit(decimal depositAmount);
     }
 }
